Enable manual reads
Allows non-blocking reads by breaking out a getFullLuminosityManual function. This function assumes that enable has been called sometime before and the user has taken care of checking the timing so that the sample is read to be read.

It performs only the operation of reading the registers and returning a uint32_t in the same format as getFullLuminosity().